+++ crates/organization/src/lib.rs
@@ -5,6 +5,8 @@ pub mod dto;
 pub mod migrations;
 pub mod models;
 pub mod module;
+#[cfg(feature = "notification")]
+mod notification;
 pub mod public;
 pub mod repositories;
 pub mod routes;